🚚 moved: functionality of /auto merged into /batch + added "advanced" filters to /batch + added button on /batch advanced mode to see the pattern results before proceeding + added option on /batch to watch any new threads created within the parent channel - /auto command shows depracation warning and directs user to /batch
✨ Added: /channel sub-command
+ /add: adds a channel to channel watch list without any filters
+ /remove: removes a channel from the channel watch list

🐛 Fixed: threads added with /batch missing first ressurection cycle
Previously, threads added via /batch were incorrectly added with the current date as the latest activity timestamp.
This left some users with threads that would stay stale for a week. This happened when the bot could not find the last message in the thread.
Threads added with /batch will now be checked next ressurection cycle.

💬 Text update: Added notice about delay in thread ressurection
a small notice now appears in the result embed that it might take upwards of an hour for the affected threads to be ressurected